design: 마이페이지 변경된 UI 디자인 적용
