+++ src/components/notice/form/NoticeForm.tsx
@@ -1,7 +1,5 @@
 import { useEffect } from 'react';
 
-import { useNavigate } from 'react-router-dom';
-
 import { useLocation, useNavigate } from 'react-router-dom';
 
 import {
