+++ src/pages/MyPage.tsx
@@ -9,10 +9,9 @@ import Header from '@/layouts/Header';
 import MainView from '@/layouts/MainView';
 
 import Title from '@/components/common/Title';
-import ScrollContainer from '@/components/common/container/ScrollContainer';
 import Faq from '@/components/faq/Faq';
-import FavoritePostCard from '@/components/user/FavoritePostCard';
-import PostAndCommentCollection from '@/components/user/PostAndCommentCollection';
+import ApplicationListOfSessionsCard from '@/components/mypage/ApplicationListOfSessionsCard';
+import MyCollection from '@/components/mypage/MyCollection';
 import UserNameImageCard from '@/components/user/UserNameImageCard';
 
 export default function MyPage() {
@@ -28,18 +27,21 @@ export default function MyPage() {
     },
     {
       label: '소속 캠퍼스',
-      value: campusList.map(({ campusName }) => campusName).join(', '),
+      value: (
+        <span>{campusList.map(({ campusName }) => campusName).join(', ')}</span>
+      ),
     },
     {
       label: '소속 교육과정',
       value: (
-        <ul className="flex-1">
+        <ul>
           {courseList.map(({ courseTitle }) => (
-            <span key={courseTitle} className="mb-1 block">
-              {courseTitle.length > 25
-                ? `${courseTitle.slice(0, 25)}...`
-                : courseTitle}
-            </span>
+            <li
+              key={courseTitle}
+              className="block max-w-60 overflow-hidden truncate"
+            >
+              {courseTitle}
+            </li>
           ))}
         </ul>
       ),
@@ -52,39 +54,29 @@ export default function MyPage() {
     <MainView>
       <Header title="마이페이지" />
 
-      <section className="mb-16 flex gap-4">
+      <section className="mb-16 grid grid-cols-[280px,430px,450px] gap-x-4">
         <UserNameImageCard />
 
-        <ul className="flex min-w-[25%] max-w-[500px] flex-col justify-between rounded-xl bg-vividGreen1 px-6 py-4">
+        <ul className="flex flex-col justify-between rounded-xl bg-vividGreen1 px-6 py-4 shadow-card">
           {userInfoList.map(({ value, label }) => (
             <li
               key={label}
-              className="flex items-start justify-between py-2.5 text-end font-medium text-white"
+              className="borer-red-500 flex items-start justify-between py-2.5 font-medium text-white"
             >
-              <span className="mr-5 font-medium text-vividGreen3">
+              <span className="mr-3 whitespace-nowrap font-medium text-vividGreen3">
                 {label}:
               </span>
               {value}
             </li>
           ))}
         </ul>
+
+        <ApplicationListOfSessionsCard />
       </section>
 
       <section className="mb-16">
         <Title title={`${name}님이 작성한 글 모음`} className="mb-[10px]" />
-        <PostAndCommentCollection />
-      </section>
-
-      <section className="mb-16 w-full flex-1">
-        <div className="mb-[10px] flex items-center justify-between">
-          <Title title={`${name}님이 찜한 글 모음`} />
-        </div>
-
-        <ScrollContainer>
-          {[1, 2, 3, 4, 5, 6].map(card => (
-            <FavoritePostCard key={card} />
-          ))}
-        </ScrollContainer>
+        <MyCollection />
       </section>
 
       <section>
